+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.0.192 (Jul 30, 2026)
+* Added support for `nullstone run` for AWS Lambda apps.
+* Fixed `nullstone exec` panic when workspace is not supported.
+ 
 # 0.0.191 (Jul 13, 2026)
 * Fixed GCP composer deployer updating environment variables.